+++ dst/app/routes/actions.js
@@ -53,9 +53,7 @@ actionsRouter.get('/moneyTransfer', permitScopes_1.default(['admin']), validator
             // tslint:disable-next-line:no-magic-numbers
             limit: (req.query.limit !== undefined) ? Math.min(req.query.limit, 100) : 100, page: (req.query.page !== undefined) ? Math.max(req.query.page, 1) : 1 });
         const actionRepo = new pecorino.repository.Action(mongoose.connection);
-        const totalCount = yield actionRepo.countTransferActions(searchConditions);
         const actions = yield actionRepo.searchTransferActions(searchConditions);
-        res.set('X-Total-Count', totalCount.toString());
         res.json(actions);
     }
     catch (error) {