+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Changed
 
+- 各リソース検索からX-Total-Countを削除
+
 ### Deprecated
 
 ### Removed